pyproject.toml
[tool.poetry]
name = "projects"
version = "0.1.0"
description = ""
authors = ["rajat <rajat@supplycompass.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.1"
djangorestframework = "^3.14.0"
black = "^23.3.0"
django-extensions = "^3.2.1"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

